Rename FPS plugin to Performance, add javadoc
